+++ src/app/core/http/invites.service.ts
@@ -2,9 +2,9 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { AuthenticationService } from '@app/authentication/authentication.service';
-import { GenericResponse } from '@shared/models/GenericResponse';
+import { GenericResponse } from '@app/models/GenericResponse';
 import { environment } from '@env';
-import { GenericPageableResponse } from '@shared/models/GenericPageableResponse';
+import { GenericPageableResponse } from '@app/models/GenericPageableResponse';
 
 @Injectable({
 	providedIn: 'root'
