+++ src/app/core/models/AuthSession.ts
