+++ src/app/core/models/AuthenticatedObject.ts
