+++ src/app/core/models/GenericResponse.ts
@@ -1,0 +1,6 @@
+export class GenericResponse<T> {
+	record: T;
+	records: Array<T>;
+	status: string;
+	message: string;
+}
