+++ src/app/layout/menu/menu-layout.component.html
@@ -9,7 +9,7 @@
 		<i class="fal fa-bars"></i>
 	</button>
 	<div
-		*ngIf="!organizationLogoUrl"
+		*ngIf="!user?.organization.avatar"
 		class="logo-container float-left"
 		[ngClass]="{ 'mx-auto': isHandset$ | async }"
 	>
@@ -21,12 +21,12 @@
 		/>
 	</div>
 	<div
-		*ngIf="!!organizationLogoUrl"
+		*ngIf="!!user?.organization.avatar"
 		class="nop-logo float-left d-flex align-items-center justify-content-center"
 		[ngClass]="{ 'mx-auto': isHandset$ | async }"
 	>
 		<img
-			src="{{ organizationLogoUrl }}"
+			src="{{ user?.organization.avatar }}"
 			alt="Logo Contabilidade"
 			class="logo unselectable mx-auto"
 		/>
@@ -35,23 +35,39 @@
 	<div class="dropdown outline-none" type="button" aria-label="Botão Logout">
 		<button
 			class="button btn-logout outline-none"
-			[ngClass]="{ 'font-large': !avatarUrl }"
+			[ngClass]="{ 'font-large': !user?.avatar }"
 			type="button"
 			data-toggle="dropdown"
 		>
 			<img
-				*ngIf="!!avatarUrl"
+				*ngIf="!!user?.avatar"
 				class="avatar"
-				src="{{ avatarUrl }}"
+				src="{{ user?.avatar }}"
 				alt="Avatar"
 			/>
-			<i *ngIf="!avatarUrl" class="fal fa-user"></i>
+			<i *ngIf="!user?.avatar" class="fal fa-user"></i>
 		</button>
 		<div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
-			<a class="dropdown-item" href="{{ profileUrl }}" target="_blank"
-				>Ver Perfil</a
+			<a
+				class="dropdown-header dropdown-item text-center py-2"
+				href="{{ profileUrl }}"
+				target="_blank"
 			>
-			<a class="dropdown-item" (click)="logout()">Sair</a>
+				<h6 class="text-dark m-0">
+					{{ user?.firstName }} {{ user?.lastName }}
+				</h6>
+				<span class="text-muted text-lowercase">
+					{{ user?.email }}
+				</span>
+			</a>
+			<a class="dropdown-item" href="{{ profileUrl }}" target="_blank">
+				<i class="fal fa-sliders-h"></i>
+				<span>Opções do usuário</span>
+			</a>
+			<a class="dropdown-item" (click)="logout()">
+				<i class="fal fa-sign-out-alt"></i>
+				<span>Sair</span>
+			</a>
 		</div>
 	</div>
 </div>
@@ -70,9 +86,7 @@
 			<div *ngFor="let menuItem of menuItems; let i = index">
 				<div *ngIf="!(isHandset$ | async); else mobileMenuItem">
 					<div
-						*ngIf="
-							userPermissionLevel <= menuItem.permissonLevelNeeded
-						"
+						*ngIf="user?.type <= menuItem.permissonLevelNeeded"
 						class="item outline-none w-100"
 						routerLinkActive="selected-item"
 						[routerLink]="[menuItem.path]"
@@ -97,9 +111,7 @@
 				</div>
 				<ng-template #mobileMenuItem>
 					<div
-						*ngIf="
-							userPermissionLevel <= menuItem.permissonLevelNeeded
-						"
+						*ngIf="user?.type <= menuItem.permissonLevelNeeded"
 						class="item outline-none w-100"
 						routerLinkActive="selected-item"
 						(click)="drawer.toggle()"
