+++ src/app/layout/menu/menu-layout.component.scss
@@ -153,3 +153,14 @@ i {
 	border-radius: 50px;
 	vertical-align: unset;
 }
+
+// Centraliza o texto de todos os itens do dropdown
+.dropdown-menu a {
+	// text-align: center;
+}
+
+// Padding em cada icone de cada item do dropdown do usuário
+.dropdown-item i {
+	padding-right: 1rem;
+	color: rgba($base-color, 0.8);
+}
