--- src/app/shared/models/GenericPageableResponse.ts
@@ -1,20 +1,0 @@
-export class PageInfo {
-
-  hasNext: boolean;
-  hasPrevious: boolean;
-
-  pageIndex: number;
-  pageSize: number;
-
-  totalPages: number;
-  totalElements: number;
-
-}
-
-export class GenericPageableResponse<T> {
-
-  pageInfo: PageInfo;
-
-  records: Array<T>;
-
-}
