--- src/app/shared/models/GenericResponse.ts
@@ -1,11 +1,0 @@
-export class GenericResponse<T> {
-
-  record: T;
-
-  records: Array<T>;
-
-  status: string;
-
-  message: string;
-
-}
